proyecto-triple/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
pandas
matplotlib
seaborn
Werkzeug
python-dotenv
email_validator
